go/skullsup: Added device attribute query to initialization
The number of LEDs, and addressing layouts differ between the skull and
bulb platforms. An address translation is required to yield the same
animation behavior on both platforms.

The attribute configurations pulled out of the bulb and skull platforms
are currently supported. This will be further generalized
on an as-needed basis.