firmware: Switched from USB to UART, platform and protocol changes
The DigiCDC driver was proving too unreliable and consumed a lot of
code space. This has been replaced with the DigisparkSoftSerial driver.

However, we can't receive data from SoftSerial when the NeoPixel update
routine is occurring; interrupts are disabled during this time.

To work around this, CMD_WAKE has been changed to CMD_SUMMON. This
command should now be sent from the host multiple times until it is
ACK'd. Once this ACK is received, it is safe to issue commands again.

Once CMD_REANIMATE is issued, the repeated CMD_SUMMON messages
will need to be used again to grab the demon's attention.

The "bulb" platform has been changed to "default" and the previous
"skull" platform definition has been removed. Moving forward, this
configuration consisting of two Adafruit NeoPixel sticks will
remain the default.

Finally, the number of max frames for this default platform
has been increased, thanks to some freed up space in the data section.