yocto/meta-skullsup: Include rngd service at start-up
It appears that the golang http.Client.Do() implementation (used by
skullsup-queue-server) may be blocking with getrandom(), as it has been
hanging until /proc/sys/kernel/random/entropy_avail appears to have
increased.

To address this, rng-tools is installed, which uses an rngd service
to supply entropy to the kernel via /dev/hwrng.